Parenthesize as-cast at shift LHS and named-integer index positions
Rust parses `expr as Ty << rhs` as `expr as (Ty << rhs)`, and parses
`expr as i32 ... as usize` left-to-right, so `INVALID as i32 - -1 as i32
as usize` becomes `(INVALID as i32) - ((-1 as i32) as usize)` instead of
`(INVALID as i32 - -1 as i32) as usize`. Both collisions broke valid
fixtures (`numeric_conversion_literals` and `named_integer_negative_index`).

Buffer the operand emission, then wrap it in parens when it ends with a
top-level `as <type>` cast. Apply this to the shift LHS path
(`writeConstShiftLeftOperandForResult`) and the named-integer-as-usize
index path (`writeNamedIntegerExpressionAsUsize`).

+++ tests/numeric_conversion_literals/main.rs
@@ -2,11 +2,11 @@ use std::cell::{RefCell};
 use std::rc::{Rc};
 
 pub fn byte_bit(i: Rc<RefCell<Option<i32>>>) -> u8 {
-    (*Rc::new(RefCell::new(Some(1 as u8))).borrow().as_ref().unwrap()) as u8 << ((*i.borrow().as_ref().unwrap()) % 8)
+    ((*Rc::new(RefCell::new(Some(1 as u8))).borrow().as_ref().unwrap()) as u8) << ((*i.borrow().as_ref().unwrap()) % 8)
 }
 
 pub fn uint64_mask(i: Rc<RefCell<Option<u32>>>) -> u64 {
-    (*Rc::new(RefCell::new(Some(1 as u64))).borrow().as_ref().unwrap()) as u64 << (*i.borrow().as_ref().unwrap())
+    ((*Rc::new(RefCell::new(Some(1 as u64))).borrow().as_ref().unwrap()) as u64) << (*i.borrow().as_ref().unwrap())
 }
 
 pub fn byte_from_expr(v: Rc<RefCell<Option<u8>>>) -> u8 {
@@ -15,6 +15,6 @@ pub fn byte_from_expr(v: Rc<RefCell<Option<u8>>>) -> u8 {
 
 fn main() {
     println!("{} {}", format!("{}", "byte literal bit:".to_string()), format!("{}", byte_bit(Rc::new(RefCell::new(Some(3))))));
-    println!("{} {}", format!("{}", "uint64 literal mask:".to_string()), format!("{}", uint64_mask(Rc::new(RefCell::new(Some(5))))));
-    println!("{} {}", format!("{}", "byte expression:".to_string()), format!("{}", byte_from_expr(Rc::new(RefCell::new(Some(4))))));
+    println!("{} {}", format!("{}", "uint64 literal mask:".to_string()), format!("{}", uint64_mask(Rc::new(RefCell::new(Some(5 as u32))))));
+    println!("{} {}", format!("{}", "byte expression:".to_string()), format!("{}", byte_from_expr(Rc::new(RefCell::new(Some(4 as u8))))));
 }
